Menambahkan komponen Admin

+++ src/pages/AdminDashboard.jsx
@@ -1,12 +1,18 @@
 // frontend/src/pages/AdminDashboard.jsx
-import React, { useEffect, useState, useContext } from "react";
+import { useEffect, useState, useContext } from "react";
 import axios from "axios";
-import { URL } from "../url";
+import { IF, URL } from "../url";
 import { UserContext } from "../context/UserContext";
 import { Table, Button, Pagination } from "antd";
-import { EyeOutlined, DeleteOutlined, MenuOutlined, LogoutOutlined } from "@ant-design/icons";
+import { SearchOutlined, EyeOutlined, DeleteOutlined, MenuOutlined, LogoutOutlined } from "@ant-design/icons";
 import PostDetails from "../component/admincomponent/PostDetails";
 import AdminMenu from "../component/admincomponent/AdminMenu";
+import UserTable from "../component/admincomponent/UserTable";
+import PostTable from "../component/admincomponent/PostTable";
+import { useTheme } from '../context/ThemeContext';
+import { FaSun, FaMoon } from 'react-icons/fa';
+import '../App.css';
+import Swal from 'sweetalert2';
 
 const AdminDashboard = () => {
   const { user, setUser } = useContext(UserContext);
@@ -18,6 +24,7 @@ const AdminDashboard = () => {
   const [selectedPost, setSelectedPost] = useState(null);
   const [rowsPerPage, setRowsPerPage] = useState(10);
   const [currentPage, setCurrentPage] = useState(1);
+  const { theme, toggleTheme } = useTheme();
 
   useEffect(() => {
     if (user?.role === "admin") {
@@ -45,20 +52,64 @@ const AdminDashboard = () => {
   };
 
   const deletePost = async (id) => {
-    try {
-      await axios.delete(URL + `/api/admin/posts/${id}`, { withCredentials: true });
-      setPosts(posts.filter(post => post._id !== id));
-    } catch (err) {
-      console.log(err);
+    const result = await Swal.fire({
+      title: 'Apakah Anda yakin?',
+      text: "Anda tidak dapat mengembalikan data ini!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Ya, hapus!',
+      cancelButtonText: 'Batal'
+    });
+
+    if (result.isConfirmed) {
+      try {
+        await axios.delete(URL + `/api/admin/posts/${id}`, { withCredentials: true });
+        setPosts(posts.filter(post => post._id !== id));
+        Swal.fire({
+          title: 'Dihapus!',
+          text: 'Data telah dihapus.',
+          icon: 'success',
+          confirmButtonText: 'OK',
+          customClass: {
+            confirmButton: 'custom-swal-button'
+          }
+        });
+      } catch (err) {
+        console.log(err);
+      }
     }
   };
 
   const deleteUser = async (id) => {
-    try {
-      await axios.delete(URL + `/api/admin/users/${id}`, { withCredentials: true });
-      setUsers(users.filter(user => user._id !== id));
-    } catch (err) {
-      console.log(err);
+    const result = await Swal.fire({
+      title: 'Apakah Anda yakin?',
+      text: "Anda tidak dapat mengembalikan data ini!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Ya, hapus!',
+      cancelButtonText: 'Batal'
+    });
+
+    if (result.isConfirmed) {
+      try {
+        await axios.delete(URL + `/api/admin/users/${id}`, { withCredentials: true });
+        setUsers(users.filter(user => user._id !== id));
+        Swal.fire({
+          title: 'Dihapus!',
+          text: 'Data telah dihapus.',
+          icon: 'success',
+          confirmButtonText: 'OK',
+          customClass: {
+            confirmButton: 'custom-swal-button'
+          }
+        });
+      } catch (err) {
+        console.log(err);
+      }
     }
   };
 
@@ -91,123 +142,6 @@ const AdminDashboard = () => {
     }
   };
 
-  const columnsPosts = [
-    {
-      title: 'No',
-      dataIndex: 'index',
-      key: 'index',
-      render: (text, record, index) => index + 1,
-    },
-    {
-      title: 'Gambar',
-      dataIndex: 'photo',
-      key: 'photo',
-      render: (text) => <img src={text} alt="No Image" className="w-16 h-16 object-cover" />,
-    },
-    {
-      title: 'Judul',
-      dataIndex: 'title',
-      key: 'title',
-      sorter: (a, b) => a.title.localeCompare(b.title),
-    },
-    {
-      title: 'Deskripsi',
-      dataIndex: 'desc',
-      key: 'desc',
-      width: 400,
-      className: 'hidden md:table-cell',
-      render: (text) => (
-        <div style={{ 
-          whiteSpace: 'normal', 
-          overflow: 'hidden', 
-          textOverflow: 'ellipsis', 
-          width: '400px', 
-          maxWidth: '100%', 
-          boxSizing: 'border-box'
-        }}>
-          {text.length > 300 ? text.substring(0, 300) + '...' : text}
-        </div>
-      ),
-    },
-    {
-      title: 'Jenis Laporan',
-      dataIndex: 'reportType',
-      key: 'reportType',
-      className: 'hidden md:table-cell',
-      sorter: (a, b) => {
-        const order = { 'pencari': 1, 'penemu': 2 };
-        return order[a.reportType] - order[b.reportType];
-      },
-    },
-    {
-      title: 'Diposting',
-      dataIndex: 'username',
-      key: 'username',
-      className: 'hidden md:table-cell',
-      sorter: (a, b) => a.username.localeCompare(b.username),
-    },
-    {
-      title: 'Kontak',
-      dataIndex: 'contactNo',
-      key: 'contactNo',
-      className: 'hidden md:table-cell',
-      sorter: (a, b) => a.contactNo.localeCompare(b.contactNo),
-    },
-    {
-      title: 'Aksi',
-      key: 'action',
-      align: 'center',
-      width: 150,
-      render: (text, record) => (
-        <div className="flex justify-center space-x-2">
-          <Button 
-            onClick={() => showPostDetails(record)} 
-            icon={<EyeOutlined />} 
-            className="flex items-center justify-center"
-          >
-            View
-          </Button>
-          <Button 
-            onClick={() => deletePost(record._id)} 
-            icon={<DeleteOutlined />} 
-            danger 
-            className="flex items-center justify-center"
-          >
-            Delete
-          </Button>
-        </div>
-      ),
-    },
-  ];
-
-  const columnsUsers = [
-    {
-      title: 'No',
-      dataIndex: 'index',
-      key: 'index',
-      render: (text, record, index) => index + 1,
-    },
-    {
-      title: 'Username',
-      dataIndex: 'username',
-      key: 'username',
-      sorter: (a, b) => a.username.localeCompare(b.username),
-    },
-    {
-      title: 'Email',
-      dataIndex: 'email',
-      key: 'email',
-      sorter: (a, b) => a.email.localeCompare(b.email),
-    },
-    {
-      title: 'Aksi',
-      key: 'action',
-      render: (text, record) => (
-        <Button onClick={() => deleteUser(record._id)} icon={<DeleteOutlined />} danger />
-      ),
-    },
-  ];
-
   const filteredPosts = posts.filter(post =>
     post.title.toLowerCase().includes(searchTerm.toLowerCase())
   );
@@ -228,33 +162,42 @@ const AdminDashboard = () => {
   }
 
   return (
-    <div className="flex">
-      <div className="flex-1 p-4">
-        <div className="flex justify-between items-center md:mx-16 lg:mx-64 mb-4">
-          <Button onClick={() => setMenuOpen(!menuOpen)} icon={<MenuOutlined />} />
+    <div className={`${theme === 'dark' ? 'dark-mode' : ''} ${theme === 'dark' ? 'bg-gray-900 text-white' : 'bg-white text-black'} min-h-screen`}>
+      <div className="flex-1 p-4 mx-4 md:mx-16 lg:mx-32">
+        <div className="flex justify-between items-center mb-4 relative">
+          <div className="relative">
+            <Button onClick={() => setMenuOpen(!menuOpen)} icon={<MenuOutlined />} />
+            {menuOpen && (
+              <AdminMenu setView={setView} setMenuOpen={setMenuOpen} />
+            )}
+          </div>
           <h1 className="hidden md:block text-3xl font-bold">Admin Dashboard</h1>
-          <Button onClick={logout} type="primary" danger className="flex items-center justify-center">
-            Logout <LogoutOutlined />
-          </Button>
+          <div className="flex items-center space-x-2">
+            <Button onClick={toggleTheme} className="flex items-center justify-center">
+              {theme === 'light' ? <FaMoon style={{ fontSize: '24px' }} /> : <FaSun style={{ fontSize: '24px' }} />}
+            </Button>
+            <Button onClick={logout} type="primary" danger className="flex items-center justify-center">
+              Logout <LogoutOutlined />
+            </Button>
+          </div>
         </div>
 
-        {menuOpen && (
-          <AdminMenu setView={setView} setMenuOpen={setMenuOpen} />
-        )}
-
-        <div className="bg-white shadow-lg rounded-lg p-4 md:mx-16 lg:mx-64">
+        <div className={`shadow-lg rounded-lg p-4 ${theme === 'dark' ? 'bg-gray-800' : 'bg-white'}`} style={{ width: '100%', overflowX: 'auto' }}>
           <div className="flex flex-col md:flex-row justify-between items-start md:items-center mb-4">
             <h2 className="text-2xl font-semibold mb-2 md:mb-0">
-              {view === "posts" ? "User Posts" : "User Management"}
+              {view === "posts" ? "Post Management" : "User Management"}
             </h2>
             <div className="flex space-x-2 ml-auto">
-              <input
-                type="text"
-                placeholder="Search here..."
-                className="border rounded px-2 py-1"
-                value={searchTerm}
-                onChange={(e) => setSearchTerm(e.target.value)}
-              />
+              <div className="flex items-center border rounded px-2 py-1">
+                <SearchOutlined className="mr-2" />
+                <input
+                  type="text"
+                  placeholder="Search here..."
+                  className="flex-1 border-none outline-none"
+                  value={searchTerm}
+                  onChange={(e) => setSearchTerm(e.target.value)}
+                />
+              </div>
               <select
                 value={rowsPerPage}
                 onChange={(e) => setRowsPerPage(Number(e.target.value))}
@@ -267,39 +210,20 @@ const AdminDashboard = () => {
             </div>
           </div>
 
+          <div className="hide-scrollbar" style={{ overflowX: 'true' }}>
           {view === "posts" ? (
-            <>
-              <Table
-                columns={columnsPosts}
-                dataSource={paginatedPosts}
-                rowKey="_id"
-                pagination={false}
-              />
-              <Pagination
-                current={currentPage}
-                pageSize={rowsPerPage}
-                total={filteredPosts.length}
-                onChange={handlePageChange}
-                className="mt-4 text-center"
-              />
-            </>
+            <PostTable posts={paginatedPosts} theme={theme} showPostDetails={showPostDetails} deletePost={deletePost} />
           ) : (
-            <>
-              <Table
-                columns={columnsUsers}
-                dataSource={paginatedUsers}
-                rowKey="_id"
-                pagination={false}
-              />
-              <Pagination
-                current={currentPage}
-                pageSize={rowsPerPage}
-                total={filteredUsers.length}
-                onChange={handlePageChange}
-                className="mt-4 text-center"
-              />
-            </>
+            <UserTable users={paginatedUsers} theme={theme} deleteUser={deleteUser} />
           )}
+          <Pagination
+            current={currentPage}
+            pageSize={rowsPerPage}
+            total={view === "posts" ? filteredPosts.length : filteredUsers.length}
+            onChange={handlePageChange}
+            className="mt-4 text-center"
+          />
+          </div>
         </div>
 
         {selectedPost && (
